refactor(user) : #52 하단바 생성 및 위젯 분리
- 하단바를 결제 페이지에서도 사용하기 위해 위젯으로 따로 분리
- page를 받을 때는 타입 Widget으로 지정
