+++ lib/user/screens/tossPayments_screen.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class TosspaymentsPage extends StatelessWidget {
+  const TosspaymentsPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('토스페이먼츠'),
+      ),
+    );
+  }
+}
